# openapi.yaml
openapi: 3.0.0
paths:
...
  /api/v1/messages/{messageId}/reactions:
    get:
      operationId: ReactionController_getDmReactions
      parameters:
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Reactions retrieved successfully
        "403":
          description: Forbidden
        "404":
          description: Message not found
      security: &a13
        - bearer: []
      summary: Get DM reactions
      tags: &a14
        - Reactions
    post:
      operationId: ReactionController_addDmReaction
      parameters:
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/AddUnicodeReactionRequestDTO"
                - $ref: "#/components/schemas/AddCustomReactionRequestDTO"
      responses:
        "201":
          description: Reaction added
        "400":
          description: Invalid emoji or limit reached
        "403":
          description: Forbidden
        "404":
          description: Message not found
      security: *a13
      summary: Add reaction to DM
      tags: *a14
    delete:
      operationId: ReactionController_removeDmReaction
      parameters:
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RemoveUnicodeReactionRequestDTO"
                - $ref: "#/components/schemas/RemoveCustomReactionRequestDTO"
      responses:
        "200":
          description: Reaction removed
        "403":
          description: Forbidden
        "404":
          description: Message not found
      security: *a13
      summary: Remove reaction from DM
      tags: *a14
  /api/v1/servers/{serverId}/channels/{channelId}/messages/{messageId}/reactions:
    post:
      operationId: ReactionController_addServerReaction
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/AddUnicodeReactionRequestDTO"
                - $ref: "#/components/schemas/AddCustomReactionRequestDTO"
      responses:
        "201":
          description: Reaction added
        "400":
          description: Invalid emoji or limit reached
        "403":
          description: Forbidden
        "404":
          description: Message or channel not found
      security: *a13
      summary: Add reaction to server message
      tags: *a14
    delete:
      operationId: ReactionController_removeServerReaction
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/RemoveUnicodeReactionRequestDTO"
                - $ref: "#/components/schemas/RemoveCustomReactionRequestDTO"
      responses:
        "200":
          description: Reaction removed
        "403":
          description: Forbidden
        "404":
          description: Message or channel not found
      security: *a13
      summary: Remove reaction from server message
      tags: *a14
    get:
      operationId: ReactionController_getServerReactions
      parameters:
        - name: serverId
          required: true
          in: path
          schema:
            type: string
        - name: channelId
          required: true
          in: path
          schema:
            type: string
        - name: messageId
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Reactions retrieved successfully
        "403":
          description: Forbidden
        "404":
          description: Not found
      security: *a13
      summary: Get server reactions
      tags: *a14
...
info:
  title: Serchat API
  description: The Serchat API description
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...